+++ src/main/java/com/clickhouse/kafka/connect/sink/ClickHouseSinkConfig.java
@@ -8,6 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.charset.StandardCharsets;
+import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -43,6 +44,7 @@ public class ClickHouseSinkConfig {
     public static final String ENABLE_DB_TOPIC_SPLIT = "enableDbTopicSplit";
     public static final String DB_TOPIC_SPLIT_CHAR = "dbTopicSplitChar";
     public static final String KEEPER_ON_CLUSTER = "keeperOnCluster";
+    public static final String DATE_TIME_FORMAT = "dateTimeFormats";
 
     public static final int MILLI_IN_A_SEC = 1000;
     private static final String databaseDefault = "default";
@@ -86,6 +88,7 @@ public class ClickHouseSinkConfig {
     private final boolean enableDbTopicSplit;
     private final String dbTopicSplitChar;
     private final String keeperOnCluster;
+    private final Map<String, DateTimeFormatter> dateTimeFormats;
 
     public enum InsertFormats {
         NONE,
@@ -245,6 +248,17 @@ public class ClickHouseSinkConfig {
         this.dbTopicSplitChar = props.getOrDefault(DB_TOPIC_SPLIT_CHAR, "");
         this.keeperOnCluster = props.getOrDefault(KEEPER_ON_CLUSTER, "");
         this.bypassRowBinary = Boolean.parseBoolean(props.getOrDefault("bypassRowBinary", "false"));
+        this.dateTimeFormats = new HashMap<>();
+        String dateTimeFormatsString = props.getOrDefault(DATE_TIME_FORMAT, "").trim();
+        if (!dateTimeFormatsString.isBlank()) {
+            String [] stringSplit = dateTimeFormatsString.split(";");
+            for (String topicToDateTimeFormat: stringSplit) {
+                String [] propSplit = topicToDateTimeFormat.trim().split("=");
+                if ( propSplit.length == 2 ) {
+                    dateTimeFormats.put(propSplit[0].trim(), DateTimeFormatter.ofPattern(propSplit[1].trim()));
+                }
+            }
+        }
 
         LOGGER.debug("ClickHouseSinkConfig: hostname: {}, port: {}, database: {}, username: {}, sslEnabled: {}, timeout: {}, retry: {}, exactlyOnce: {}",
                 hostname, port, database, username, sslEnabled, timeout, retry, exactlyOnce);
@@ -521,6 +535,15 @@ public class ClickHouseSinkConfig {
                 ++orderInGroup,
                 ConfigDef.Width.SHORT,
                 "Bypass RowBinary format.");
+        configDef.define(DATE_TIME_FORMAT,
+                ConfigDef.Type.LIST,
+                "",
+                ConfigDef.Importance.LOW,
+                "Date time formats for parsing date time fields (e.g. 'someDateField=yyyy-MM-dd HH:mm:ss.SSSSSSSSS'). default: ''",
+                group,
+                ++orderInGroup,
+                ConfigDef.Width.SHORT,
+                "Date time formats.");
         return configDef;
     }
 }